+++ stable_baselines/__init__.py
@@ -1,3 +1,4 @@
 from stable_baselines.td3 import TD3
+from stable_baselines.ppo import PPO
 
 __version__ = "3.0.0a0"