+++ stable_baselines/ppo/__init__.py
@@ -1,0 +1,2 @@
+from stable_baselines.ppo.ppo import PPO
+from stable_baselines.ppo.policies import MlpPolicy